+++ src/main/java/com/example/sistemadeingresssos/repositories/IngressoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IngressoRepository extends JpaRepository<Ingresso, Long> {
+public interface IngressoRepository extends JpaRepository<Ingresso, Integer> {
 }
